% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PanelScript.R
\name{storeVar}
\alias{store}
\alias{storeV}
\alias{storeVar}
\alias{storeVariable}
\title{Store raw demographic and political variables for all unique emails in the panel}
\usage{
storeVar(var, Data, toFile, fromFile = NA, idvar = NA, source = NA,
  reportFile = NA)
}
\arguments{
\item{var}{a vector of variabels to store}

\item{Data}{a data frame of the input dataset to extract raw variabels from}

\item{toFile}{a vector of paths of the output datasets}

\item{fromFile}{a vector of paths of the input datasets}

\item{idvar}{a vector of id variables for the input and output datasets}

\item{source}{a string of source of the input dataset}

\item{reportFile}{a string path of a report file}
}
\description{
Store raw demographic and political variables for all unique emails in the panel
}
\examples{
library(plyr)
setwd("~/Dropbox (Vox Pop Labs)/Data/Panel/CanadaPanel/RawData")
parents.dir <- list.files()
fromFilePaths <- unlist(apply(as.matrix(parents.dir), 1, function(x) paste0(x, "/", list.files(x))))
Data <- list()

for (i in 1:length(fromFilePaths)) {
Data[[i]] <- read.csv(fromFilePaths[i], stringsAsFactors = FALSE)
Data[[i]]$source <- gsub("\\\\/", "-", gsub("EmailData.csv", "", fromFilePaths[i]))
if ( grepl("Toronto2014", fromFilePaths[i]) ) { Data[[i]]$province <- "Ontario"
} else if ( grepl("Alberta2015", fromFilePaths[i]) ) {
		Data[[i]]$province <- "Alberta"
}
		names(Data)[names(Data) \%in\% "province"] <- "provinceRaw"
}
varList <- c("genderRaw", "birthYearRaw", "religionRaw", "incomeRaw", "educationRaw", "selfPlacementRaw", "polInterestRaw", "voteChoiceRaw")
varList <- c("occupationRaw")
varList <- c("province", "postal_code", "riding")
toFilePaths <- paste0("../Variables/", varList, ".csv")
l_ply(Data, function(D) { storeVar(varList, D, toFilePaths, idvar = c("date", "uniqueID"), source = unique(D$source), reportFile = "../Report.csv") })
}

