% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PanelScript.R
\name{updateMaster}
\alias{Master}
\alias{master}
\alias{updateM}
\alias{updateMaster}
\alias{updatemaster}
\title{Update the master file}
\usage{
updateMaster(masterFile, fromFile, toFile, id = "email", dateRange = c(0,
  Inf), newIdOnly = FALSE, newIdFile)
}
\arguments{
\item{masterFile}{a data frame or a path of the master file}

\item{fromFile}{a list of data frames or paths of the input datasets}

\item{toFile}{a string path of the updated master file}

\item{id}{a string of variable is used as id variable. By default, id variable is email.}

\item{dateRange}{a vector of lower and upper bound of dates that are selected from the input file.}

\item{newIdOnly}{a logical variable to indicate whether or not to update only new IDs.}

\item{newIdFile}{a string of the path that contains new IDs.}
}
\description{
Update the master file
}
\examples{
masterFile <- read.csv("../Variables/email.csv", stringsAsFactors = FALSE)

fromFileList <- paste0("../Variables/", list.files("../Variables/")[!grepl("email.csv", list.files("../Variables"))])
updateMaster(masterFile, as.list(fromFileList), "../Canada master file.csv")
masterFile <- masterFile[-sample(1:nrow(masterFile), 200), ]
updateMaster(masterFile, fromFileList, "../Canada master file2.csv", newIdOnly = TRUE, newIdFile = "../Variables/email.csv")
}

